# openapi/devflow-lengkap.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevFlow — API Lengkap",
    "version": "1.0",
    "description": "Gabungan spesifikasi kelima service backend DevFlow, seluruhnya di satu berkas.\n\nDihasilkan dari controller lewat springdoc, lalu digabung — bukan ditulis tangan, jadi isinya mengikuti kode.\n\nSetiap operasi diberi tag nama service asalnya. Ambil token lewat POST /api/v1/auth/login, lalu pakai skema bearerAuth."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "api-gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "project-service"
        ],
        "operationId": "getProjectsByWorkspace",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "project-service"
        ],
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/time-entries": {
      "get": {
        "tags": [
          "project-service"
        ],
        "operationId": "getTimeEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTimeEntryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "project-service"
        ],
        "operationId": "createTimeEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TimeEntryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{id}/time-entries/{entryId}": {
      "delete": {
        "tags": [
          "project-service"
        ],
        "operationId": "deleteTimeEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "project-service"
        ],
        "operationId": "updateTimeEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TimeEntryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{id}/time-entries/user/{targetUserId}": {
      "get": {
        "tags": [
          "project-service"
        ],
        "operationId": "getUserTimeEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeEntryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/time-entries/recent": {
      "get": {
        "tags": [
          "project-service"
        ],
        "operationId": "getRecentEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeEntryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{id}/dashboard": {
      "get": {
        "tags": [
          "project-service"
        ],
        "operationId": "getDashboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDashboardResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNING",
              "IN_PROGRESS",
              "ON_HOLD",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "clientVisible": {
            "type": "boolean"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "demoUrl": {
            "type": "string"
          },
          "githubRepoOwner": {
            "type": "string"
          },
          "githubRepoName": {
            "type": "string"
          },
          "githubRepoBranch": {
            "type": "string"
          },
          "githubLinkedBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateProjectRequest": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "clientVisible": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "workspaceId"
        ]
      },
      "TimeEntryRequest": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "entryDate": {
            "type": "string",
            "format": "date"
          },
          "hours": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 24.0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "entryType": {
            "type": "string",
            "minLength": 1
          },
          "featureName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "entryDate",
          "entryType",
          "hours"
        ]
      },
      "TimeEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "taskTitle": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          },
          "entryDate": {
            "type": "string",
            "format": "date"
          },
          "hours": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "entryType": {
            "type": "string"
          },
          "featureName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageTimeEntryResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeEntryResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "MemberWorkloadDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "hoursThisWeek": {
            "type": "number"
          },
          "totalHours": {
            "type": "number"
          },
          "tasksAssigned": {
            "type": "integer",
            "format": "int64"
          },
          "tasksDone": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ProjectDashboardResponse": {
        "type": "object",
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "totalHours": {
            "type": "number"
          },
          "hoursThisWeek": {
            "type": "number"
          },
          "hoursToday": {
            "type": "number"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64"
          },
          "taskCount": {
            "type": "integer",
            "format": "int64"
          },
          "doneTaskCount": {
            "type": "integer",
            "format": "int64"
          },
          "overdueTaskCount": {
            "type": "integer",
            "format": "int64"
          },
          "recentEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeEntryResponse"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberWorkloadDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Access token dari /api/v1/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}